Convert the remaining floorf to floor.

+++ UIKit/Classes/UIColorRep.m
@@ -50,7 +50,7 @@ static void drawPatternImage(void *info, CGContextRef ctx)
     // appears flipped for reasons I don't fully understand unless I apply this counter-flip transform. If the UIView does
     // have a drawRect:, then the different way that the background color is set (UIView draws it directly into the
     // CGContext that Core Animation gives it before calling drawRect:), causes the pattern to appear right-side-up.
-    if (floorf(NSAppKitVersionNumber) != NSAppKitVersionNumber10_7) {
+    if (floor(NSAppKitVersionNumber) != NSAppKitVersionNumber10_7) {
         if (CGPointEqualToPoint(patternRect.origin, deviceRect.origin)) {
             CGContextTranslateCTM(ctx, 0, patternRect.size.height);
             CGContextScaleCTM(ctx, 1, -1);
